Add files via upload
Using the Non-linearization procedure, this program calculates the perturbative corrections of the Hydrogen atom subjected to a constant magnetic field. More information can be found at: Turbiner, A. V. (1984). The eigenvalue spectrum in quantum mechanics and the nonlinearization procedure. Soviet Physics Uspekhi, 27(9), 668.

This Mathematica's notebook is optimized to be run from the terminal/console.